Consolidate controller, move rpc package into controller - remove dangling code in pkg/server
