+++ pkg/controller/rpc/auth.go
